+++ packages/expo-ui/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from './ExpoUIModule';
+export { default as SingleChoiceSegmentedControlView } from './ExpoUIView';
+export * from './ExpoUI.types';